create config folder in dockerfile
I am running TRD in kubernetes and injecting the config file from a
configmap.
Since we changed the user to `trd`, I need this folder to exist and
owned by TRD, otherwise kubernetes will create it for me and have it
owned by `root`, so TRD can not put a lock file in it.